# openapi.yml
openapi: 3.0.3
info:
  title: buff
  description: ""
  license:
    name: ""
  version: 0.1.0
paths:
  /api/admin/blacklist:
    get:
      tags:
        - handlers::admin::blacklist
      operationId: full_blacklist
      responses:
        "200":
          description: Blacklist was successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  format: int64
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_admin: []
    post:
      tags:
        - handlers::admin::blacklist
      operationId: blacklist_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlacklistUserRequest"
        required: true
      responses:
        "204":
          description: User was successfully blacklisted
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "404":
          description: User was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "409":
          description: User has already been blacklisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_admin: []
    delete:
      tags:
        - handlers::admin::blacklist
      operationId: unblacklist_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnblacklistUserRequest"
        required: true
      responses:
        "204":
          description: User was successfully unblacklisted
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "404":
          description: User was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "409":
          description: User is not blacklisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_admin: []
  /api/admin/currency:
    post:
      tags:
        - handlers::admin::currency
      operationId: create_currency
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCurrencyRequest"
        required: true
      responses:
        "201":
          description: Currency rate was successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
  /api/admin/currency/{id}:
    delete:
      tags:
        - handlers::admin::currency
      operationId: delete_currency_rate_by_id
      parameters:
        - name: id
          in: path
          description: Currency rate id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Currency rate was successfully deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "404":
          description: Currency was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_admin: []
    patch:
      tags:
        - handlers::admin::currency
      operationId: set_currency_rate_by_id
      parameters:
        - name: id
          in: path
          description: Currency rate id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetRateRequest"
        required: true
      responses:
        "204":
          description: Currency rate was successfully changed
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "404":
          description: Currency was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Details"
      security:
        - jwt_admin: []
...
components:
  schemas:
...
    BlacklistUserRequest:
      type: object
      required:
        - steam_id
      properties:
        steam_id:
          type: integer
          format: int64
...
    CreateCurrencyRequest:
      type: object
      required:
        - symbol
        - rate
      properties:
        symbol:
          type: string
        rate:
          type: string
...
    Details:
      type: object
      required:
        - details
      properties:
        details:
          type: string
...
    Order:
      type: object
      required:
        - id
        - payment_method
        - status
        - created_at
        - steam_id
        - amount
        - fixed_currency_rate
        - currency_symbol
      properties:
        id:
          type: integer
          format: int64
        payment_method:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        steam_id:
          type: integer
          format: int64
        moderator_id:
          type: integer
          format: int64
          nullable: true
        amount:
          type: string
        fixed_currency_rate:
          type: string
        currency_symbol:
          type: string
        finished_at:
          type: string
          format: date-time
          nullable: true
...
    SetRateRequest:
      type: object
      required:
        - rate
      properties:
        rate:
          type: string
...
    UnblacklistUserRequest:
      type: object
      required:
        - steam_id
      properties:
        steam_id:
          type: integer
          format: int64
...
  securitySchemes:
    jwt_admin:
      type: apiKey
      in: header
      name: X-AM-Authorization
    jwt_user:
      type: http
      scheme: bearer
      bearerFormat: JWT